# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "GeekHub COPY of Telegram API",
    "description": "Documentation for GeekHub graduation project node.js API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "host": "api.example.com",
  "basePath": "/v1",
  "schemes": [
    "https"
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "User",
      "description": "API for users in the system"
    },
    {
      "name": "File",
      "description": "API for files in the system"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/file": {
      "post": {
        "tags": [
          "File"
        ],
        "summary": "Uploads a file.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/file/{pathname}": {
      "get": {
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "pathname",
            "schema": {
              "type": "string",
              "minimum": 3
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file."
          }
        }
      }
    },
    "/user/info": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "summary": "Get user info",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": null,
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string",
              "minimum": 3
            }
          },
          {
            "in": "query",
            "name": "password",
            "schema": {
              "type": "string",
              "minimum": 8
            }
          }
        ],
        "summary": "Login user",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/definitions/User"
                    },
                    {
                      "type": "object"
                    }
                  ],
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Registration user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/definitions/UserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": null,
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "id": {
      "properties": {
        "uuid": {
          "type": "string"
        }
      }
    },
    "UserCommon": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      }
    },
    "UserDTO": {
      "allOf": [
        {
          "$ref": "#/definitions/UserCommon"
        }
      ],
      "properties": {
        "password": {
          "type": "string"
        }
      }
    },
    "User": {
      "allOf": [
        {
          "$ref": "#/definitions/UserCommon"
        },
        {
          "type": "object",
          "required": [
            "rootCause"
          ],
          "properties": {
            "password": {
              "not": null
            },
            "_id": {
              "type": "string"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time"
            }
          }
        }
      ]
    }
  }
}